Start organizing new types some
