+++ src/types/meta.ts
@@ -1,0 +1,3 @@
+import { RouteMeta } from './register'
+
+export type ToMeta<TMeta extends RouteMeta | undefined> = TMeta extends undefined ? {} : unknown extends TMeta ? {} : TMeta
